+++ src/components/pages/apps/genomic/seq-browser/dialogs/edit/genes-edit-dialog.tsx
@@ -1,13 +1,7 @@
 import { DoubleNumericalInput } from '@/components/double-numerical-input'
-import { VCenterRow } from '@/components/layout/v-center-row'
 import { FontUI } from '@/components/plot/font/font-ui'
 import { ScrollAccordion } from '@/components/shadcn/ui/themed/v2/accordion'
 import { LabelContainer } from '@/components/shadcn/ui/themed/v2/label'
-import {
-  GroupIndicatorToggle,
-  ToggleGroup,
-} from '@/components/shadcn/ui/themed/v2/toggle-group'
-import { TabIndicatorIosSelected } from '@/components/tabs/tab-indicator-ios-selected'
 import { TabIndicatorProvider } from '@/components/tabs/tab-indicator-provider'
 import { TEXT_CANCEL, TEXT_OK } from '@/consts'
 import { ColorPropRow } from '@/dialogs/color-prop-row'
@@ -26,6 +20,7 @@ import {
   type GeneView,
 } from '../../seq-browser-settings'
 import type { IGeneTrack, ITrackGroup } from '../../tracks-provider'
+import { IosGroupToggle } from './ios-group-toggle'
 
 export interface IProps extends IModalProps<{
   group: ITrackGroup
@@ -430,45 +425,26 @@ export function GenesEditDialog({ group, track, onResponse }: IProps) {
             </SelectList>
           </PropRow>
           <PropRow title="View" className="items-start">
-            <VCenterRow className="rounded-full p-0.5 overflow-hidden bg-muted/50 text-xs relative">
-              <TabIndicatorProvider>
-                <ToggleGroup
-                  value={[settings.tracks.genes.view]}
-                  onValueChange={(v, e) => {
-                    if (!v || v.length === 0) {
-                      return
-                    }
-                    const newSettings = produce(settings, (draft) => {
-                      draft.tracks.genes.view = v[0] as GeneView
-                    })
+            <TabIndicatorProvider>
+              <IosGroupToggle
+                w={5.5}
+                tabs={[
+                  { id: 'transcript', name: 'Transcripts' },
+                  { id: 'features', name: 'Features' },
+                ]}
+                value={[settings.tracks.genes.view]}
+                onValueChange={(v, e) => {
+                  if (!v || v.length === 0) {
+                    return
+                  }
+                  const newSettings = produce(settings, (draft) => {
+                    draft.tracks.genes.view = v[0] as GeneView
+                  })
 
-                    updateSettings(newSettings)
-                  }}
-                  size="toolbar"
-                  className="rounded-full overflow-hidden gap-x-px relative"
-                  variant="ios"
-                >
-                  <GroupIndicatorToggle
-                    isSelected={settings.tracks.genes.view === 'transcript'}
-                    value="transcript"
-                    className="w-22"
-                    rounded="full"
-                  >
-                    Transcripts
-                  </GroupIndicatorToggle>
-                  <GroupIndicatorToggle
-                    isSelected={settings.tracks.genes.view === 'features'}
-                    value="features"
-                    className="w-22"
-                    rounded="full"
-                  >
-                    Features
-                  </GroupIndicatorToggle>
-
-                  <TabIndicatorIosSelected defaultWidth="5.5rem" />
-                </ToggleGroup>
-              </TabIndicatorProvider>
-            </VCenterRow>
+                  updateSettings(newSettings)
+                }}
+              />
+            </TabIndicatorProvider>
           </PropRow>
         </SettingsAccordionItem>
       </ScrollAccordion>