+++ src/components/shadcn/ui/themed/v2/toggle.tsx
@@ -28,7 +28,7 @@ export const toggleVariants = cva(TOGGLE_CLS, {
         'flex flex-col gap-y-2 items-center p-2 border-2 border-border border-transparent hover:border-app-theme/25 data-pressed:shadow-xs data-pressed:border-app-theme/50 data-pressed:bg-app-theme/10',
       group:
         'hover:bg-muted/60 data-pressed:text-theme focus-visible:z-10 focus:z-10 outline-2 -outline-offset-2 outline-transparent focus-visible:outline-ring border border-border data-pressed:font-semibold',
-      ios: 'focus-visible:outline-ring data-pressed:font-semibold',
+      ios: 'data-pressed:font-semibold',
     },
     size: {
       sm: 'h-7',
@@ -48,7 +48,7 @@ export const toggleVariants = cva(TOGGLE_CLS, {
     },
     rounded: {
       none: '',
-      sm: 'rounded-xs',
+      sm: 'rounded-sm',
       default: 'rounded-theme',
       lg: 'rounded-lg',
       full: 'rounded-full',